Remove Cases Missing All Measurements

Function to remove the table rows for which all measurements of interest are non-available (NA). A particular list of measurement names can be explicitly provided or selected by a common initial pattern. The default setting removes the rows with no log-ratio available.

RemoveNACases(data, measureNames = NULL, prefix = logPrefix)

Arguments

data
A dataframe with the input measurements.

measureNames
A vector of characters with the list of measurements to be considered for missing values. If NULL (default), all measurements starting by prefix are considered.

prefix
A character string with the initial pattern to select the list of measurements. The default is given by the internal variable logPrefix. It is in effect only when measureNames = NULL.

Value

A dataframe with the same columns as the input dataframe but removing the rows with missing values for all measurements in the list.

Examples

## Read an example dataset:
dataFile <- system.file("extdata", "dataValenzuelaLamas2008.csv.gz",
                        package="zoolog")
dataExample <- utils::read.csv2(dataFile,
                                quote = "\"", na = "", header = TRUE,
                                fileEncoding = "UTF-8")
## We can observe the first lines (excluding some columns for visibility):
head(dataExample)[, -c(6:20,32:63)]

#>   Site N.inv    UE Especie    Os   GL   Bp   Dp   SD   DD   Bd   Dd BT GLc BFd
#> 1  ALP  3629  8427    avif 1 fal   NA   NA   NA   NA   NA   NA   NA NA  NA  NA
#> 2  ALP  4918 10364    bota 1 fal 54.0 31.3 30.6 28.1 26.3 27.5 20.0 NA  NA  NA
#> 3  ALP  4919 10364    bota 1 fal 54.5 27.9 31.8 26.0 22.8 25.3 19.5 NA  NA  NA
#> 4  ALP  2211  7105    bota 1 fal 47.6 25.2 25.2 20.6 23.7 17.8   NA NA  NA  NA
#> 5  ALP  1257  7093    bota 1 fal 51.1 24.1  2.7 21.5 17.0   NA   NA NA  NA  NA
#> 6  ALP  4693 10258      oc 1 fal 31.0 10.7 12.4   NA  7.4  8.9  8.0 NA  NA  NA
#>   Dl HmandM3
#> 1 NA      NA
#> 2 NA      NA
#> 3 NA      NA
#> 4 NA      NA
#> 5 NA      NA
#> 6 NA      NA

## Remove the cases not including any measurement present in the reference.
refMeasureNames <- levels(factor(reference$Combi$Measure))
refMeasureNames

#>  [1] "BFd" "BFp" "BT"  "Bd"  "Bp"  "Dd"  "Dp"  "GL"  "GLc" "GLl" "HTC" "SD" 
#> [13] "BG"  "BPC" "DC"  "DPA" "Dl"  "GB"  "GLP" "H"   "L"   "LA"  "LG"  "SLC"
#> [25] "WCM" "GLm" "LAR" "LeP" "SDO" "SH"  "GLI"

dataExamplePruned <- RemoveNACases(dataExample,
                                   measureNames = refMeasureNames)
## The first lines of the output data frame show at least one available
## measurement value in the selected list:
head(dataExamplePruned)[, -c(6:20,32:63)]

#>   Site N.inv    UE Especie    Os   GL   Bp   Dp   SD   DD   Bd   Dd BT GLc BFd
#> 1  ALP  4918 10364    bota 1 fal 54.0 31.3 30.6 28.1 26.3 27.5 20.0 NA  NA  NA
#> 2  ALP  4919 10364    bota 1 fal 54.5 27.9 31.8 26.0 22.8 25.3 19.5 NA  NA  NA
#> 3  ALP  2211  7105    bota 1 fal 47.6 25.2 25.2 20.6 23.7 17.8   NA NA  NA  NA
#> 4  ALP  1257  7093    bota 1 fal 51.1 24.1  2.7 21.5 17.0   NA   NA NA  NA  NA
#> 5  ALP  4693 10258      oc 1 fal 31.0 10.7 12.4   NA  7.4  8.9  8.0 NA  NA  NA
#> 6  ALP  1550  7014      oc 1 fal 34.1  0.0  0.0  9.2 11.1  9.4   NA NA  NA  NA
#>   Dl HmandM3
#> 1 NA      NA
#> 2 NA      NA
#> 3 NA      NA
#> 4 NA      NA
#> 5 NA      NA
#> 6 NA      NA

## If we compute first the log-ratios
dataExampleWithLogs <- LogRatios(dataExample)
## the cases not including any log-ratio can be removed with the
## default logPrefix
dataExampleWithLogsPruned <- RemoveNACases(dataExampleWithLogs)
head(dataExampleWithLogsPruned)[, -c(6:20,32:63)]

#>   Site N.inv    UE Especie       Os   GL   Bp   Dp   SD   DD   Bd Dd BT GLc BFd
#> 1  ALP  3016  8525    ceel     1fal 50.8 17.9 23.4 15.1 17.8 15.2 NA NA  NA  NA
#> 2  ALP  2907  8068    ceel     1fal 52.3 20.1 22.6 15.4  1.7  1.6 NA NA  NA  NA
#> 3  ALP  2906  8068    ceel     1fal 52.9 17.2   NA 14.9 17.7 15.5 NA NA  NA  NA
#> 4  ALP  3453 10410    ovar 1fal ant 27.1  9.9 12.3 17.9  9.0  9.0 NA NA  NA  NA
#> 5  ALP  3455 10410    ovar 1fal ant 27.6  9.6 12.2  7.6  8.9  8.3 NA NA  NA  NA
#> 6  ALP  3524 10122    ovar 1fal ant 27.7  9.2 11.5  7.5  9.4  8.1 NA NA  NA  NA
#>   Dl HmandM3       logBd logBFd logBG       logBp logBPC logBT logDd logDl
#> 1 NA      NA          NA     NA    NA -0.07348283     NA    NA    NA    NA
#> 2 NA      NA          NA     NA    NA -0.02313980     NA    NA    NA    NA
#> 3 NA      NA          NA     NA    NA -0.09080741     NA    NA    NA    NA
#> 4 NA      NA -0.08911977     NA    NA -0.07991177     NA    NA    NA    NA
#> 5 NA      NA -0.12428419     NA    NA -0.09327573     NA    NA    NA    NA
#> 6 NA      NA -0.13487726     NA    NA -0.11175913     NA    NA    NA    NA
#>         logDp logDPA logGB       logGL logGLc logGLP logH logL logLA logLAR
#> 1          NA     NA    NA          NA     NA     NA   NA   NA    NA     NA
#> 2          NA     NA    NA          NA     NA     NA   NA   NA    NA     NA
#> 3          NA     NA    NA          NA     NA     NA   NA   NA    NA     NA
#> 4 -0.07265930     NA    NA -0.10786052     NA     NA   NA   NA    NA     NA
#> 5 -0.07620458     NA    NA -0.09992073     NA     NA   NA   NA    NA     NA
#> 6 -0.10186657     NA    NA -0.09835005     NA     NA   NA   NA    NA     NA
#>   logLG      logSD logSH logSLC
#> 1    NA         NA    NA     NA
#> 2    NA         NA    NA     NA
#> 3    NA         NA    NA     NA
#> 4    NA  0.2629585    NA     NA
#> 5    NA -0.1090810    NA     NA
#> 6    NA -0.1148333    NA     NA