Example dataset

The dataset provided as an example originates from (Valenzuela-Lamas 2008) . The dataset is written in Catalan (with the exception of some headings to ease the understanding about what is in there).

Format

The dataset is provided in the zoolog extdata folder as a file in semicolon separate format but compressed with gzip to reduce its weight:

dataValenzuelaLamas2008.csv.gz

The file is provided in UTF-8 encoding. The file encoding is relevant because the dataset contains accents and special characters that needs to be correctly displayed. It can be directly open by utils::read.csv2, provided that the correct file format is set (see examples below).

Every row of the data.frame refers to one individual bone fragment unless otherwise stated in the Observations field ("Observacions").

All the measurements are expressed in millimetres and were obtained with a manual calliper.

The main headings in the database are:

Site

The faunal remains from three Iron Age archaeological sites were recorded (ALP = Alorda Park, TFC = Turó de la Font de la Canya, OLD = Olèrdola).

N inv

A correlative number for each fragment.

UE

Refers to the Stratigraphic Unit (SU in English).

Especie

Refers to the species.

Os

Refers to the skeletal element.

Fragment

Refers to the preserved part in the vertical axis (distal, proximal, diaphysis, etc.).

Lat

Bone laterality: right (d) or left (e).

Vora

Refers to the preserved part in relation to the circumference (c), or a vertically, transversally and obliquely fragmented (sto).

Fract

Refers to fracture during field excavation or lab work.

Tafo

Refers to anthropic and post-depositional alterations.

Grau

Refers to degree of bone alteration in a scale from 0 (no alteration) to 4 (diaphysis completely altered).

Epif

Degree of fusion: s= fused, ns= unfused, ec = fusion visible. Also tooth wear is recorded here following (Gardeisen 1997) .

Sexe

Sex: male (masc) / female (fem).

Traces

Refers to butchery marks. It may also include other observations.

Observacions

Observations.

Recinte

Refers to the number of silo structure (e.g. SJ8) or the room (e.g. AB) from which the material originates.

TPQ

Absolute chronology in Terminus Post Quem.

TAQ

Absolute chronology in Terminus Ante Quem.

Period

Chronological phasing.

Capsa

Box number that contains the item.

Measurement codes

The nomenclature follows (Von den Driesch 1976) .

References

Gardeisen A (1997). “Exploitation des prélèvements et fichiers de spécialité (PRL, FAUNE, OS).” Lattara, 10, 251--278.

Valenzuela-Lamas S (2008). Alimentació i ramaderia al Penedès durant la protohistòria (segles VII-III aC). Societat Catalana d'Arqueologia, pdf version (5a edicio) here: http://www.scarqueologia.com/?page_id=10.

Von den Driesch A (1976). A guide to the measurement of animal bones from archaeological sites: as developed by the Institut für Palaeoanatomie, Domestikationsforschung und Geschichte der Tiermedizin of the University of Munich, volume 1. Peabody Museum Press.

Examples

dataFile <- system.file("extdata", "dataValenzuelaLamas2008.csv.gz",
                        package="zoolog")
dataExample <- utils::read.csv2(dataFile,
                               quote = "\"", na = "", header = TRUE,
                               fileEncoding = "UTF-8")