Taxonomy hierarchy for zoolog

The taxonomy hierarchy for all taxa included in the osteometrical references of the package zoolog. This is used allow the users to group the taxa by any taxonomical category from genus to Family. See GetTaxaIn.

zoologTaxonomy

Format

The taxonomy is given as a data.frame with columns for Taxon, Genus, Tribe, Subfamily, and Family. Each row lists the information for one taxon:

Taxon	Genus	Tribe	Subfamily	Family

Bos taurus	Bos	Bovini	Bovinae	Bovidae

Bos primigenius	Bos	Bovini	Bovinae	Bovidae

Ovis aries	Ovis	Caprini	Caprinae	Bovidae

Ovis orientalis	Ovis	Caprini	Caprinae	Bovidae

Capra hircus	Capra	Caprini	Caprinae	Bovidae

Capra aegagrus	Capra	Caprini	Caprinae	Bovidae

Gazella gazella	Gazella	Antilopini	Antilopinae	Bovidae

Sus domesticus	Sus	Suini	Suinae	Suidae

Sus scrofa	Sus	Suini	Suinae	Suidae

Cervus elaphus	Cervus	Cervini	Cervinae	Cervidae

Dama mesopotamica	Dama	Cervini	Cervinae	Cervidae

Equus asinus	Equus	Equini	Equinae	Equidae

Equus caballus	Equus	Equini	Equinae	Equidae

Oryctolagus cuniculus	Oryctolagus			Leporidae

File Structure

zoologTaxonomy is an exported variable automatically loaded in memory. In addition, the csv source file zoologTaxonomy.csv generating it is included in the zoolog extdata folder.