Condense Measure Log-Ratios

This function condenses the calculated log ratio values into a reduced number of features by grouping log ratio values and selecting or calculating a feature value. By default the selected groups represent each a single dimension, i.e. Length and Width. Only one feature is extracted per group. Currently, two methods are possible: priority (default) or average.

CondenseLogs(
  data,
  grouping = list(Length = c("GL", "GLl", "HTC"), Width = c("Bd", "BT", "Bp", "SD")),
  method = "priority"
)

Arguments

data
A dataframe with the input measurements.

grouping
A list of named character vectors. The list includes a vector per selected group. Each vector gives the group of measurements in order of priority. By default the groups are Length = c("GL", "GLl", "HTC") and Width = c("Bd", "BT", "Bp", "SD"). The order is irrelevant for method = "average".

method
Character string indicating which method to use for extracting the condensed features. Currently accepted methods: "priority" (default) and "average".

Value

A dataframe including the input dataframe and additional columns, one for each extracted condensed feature, with the corresponding name given in grouping.

Details

This operation is motivated by two circumstances. First, not all measurements are available for every bone specimen, which obstructs their direct comparison and statistical analysis. Second, several measurements can be strongly correlated (e.g. SD and Bd both represent bone width). Thus, considering them as independent would produce an over-representation of bone remains with more measurements per axis. Condensing each group of measurements into a single feature (e.g. one measure per axis) palliates both problems.

Observe that an important property of the log-ratios from a reference is that it makes the different measures comparable. For instance, a bone with a width doubling the reference will produce the same log-ratio (log(2)) for Bd, BT, Bp, and SD. In contrast, the absolute measures are not directly comparable.

The measurement names in the grouping list are given without the logPrefix. But the selection is made from the log-ratios.

The default method is "priority", which selects the first available measure log-ratio in each group. The method "average" extracts the mean per group, ignoring the non-available measures. We provide the following by-default group and prioritization: For lengths, the first option is GL, then HTC. For widths, the order of priority is: Bd, BT, Bp, SD. This order maximises the robustness and reliability of the measurements, as priority is given to the most abundant, more replicable, and less age dependent measurements.

This method was first used in: Trentacoste, A., Nieto-Espinet, A., & Valenzuela-Lamas, S. (2018). Pre-Roman improvements to agricultural production: Evidence from livestock husbandry in late prehistoric Italy. PloS one, 13(12), e0208109.

Examples

## Read an example dataset:
dataFile <- system.file("extdata", "dataValenzuelaLamas2008.csv.gz",
                        package="zoolog")
dataExample <- utils::read.csv2(dataFile,
                                quote = "\"", na = "", header = TRUE,
                                fileEncoding = "UTF-8")
## Compute the log-ratios:
dataExampleWithLogs <- LogRatios(dataExample)
## We can observe the first lines (excluding some columns for visibility):
head(dataExampleWithLogs)[, -c(6:20,32:63)]

#>   Site N.inv    UE Especie    Os   GL   Bp   Dp   SD   DD   Bd   Dd BT GLc BFd
#> 1  ALP  3629  8427    avif 1 fal   NA   NA   NA   NA   NA   NA   NA NA  NA  NA
#> 2  ALP  4918 10364    bota 1 fal 54.0 31.3 30.6 28.1 26.3 27.5 20.0 NA  NA  NA
#> 3  ALP  4919 10364    bota 1 fal 54.5 27.9 31.8 26.0 22.8 25.3 19.5 NA  NA  NA
#> 4  ALP  2211  7105    bota 1 fal 47.6 25.2 25.2 20.6 23.7 17.8   NA NA  NA  NA
#> 5  ALP  1257  7093    bota 1 fal 51.1 24.1  2.7 21.5 17.0   NA   NA NA  NA  NA
#> 6  ALP  4693 10258      oc 1 fal 31.0 10.7 12.4   NA  7.4  8.9  8.0 NA  NA  NA
#>   Dl HmandM3 logBd logBFd logBG logBp logBPC logBT logDd logDl logDp logDPA
#> 1 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 2 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 3 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 4 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 5 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 6 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#>   logGB logGL logGLc logGLP logH logL logLA logLAR logLG logSD logSH logSLC
#> 1    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 2    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 3    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 4    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 5    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 6    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA

## Extract the default condensed features with the default "priority" method:
dataExampleWithSummary <- CondenseLogs(dataExampleWithLogs)
head(dataExampleWithSummary)[, -c(6:20,32:63)]

#>   Site N.inv    UE Especie    Os   GL   Bp   Dp   SD   DD   Bd   Dd BT GLc BFd
#> 1  ALP  3629  8427    avif 1 fal   NA   NA   NA   NA   NA   NA   NA NA  NA  NA
#> 2  ALP  4918 10364    bota 1 fal 54.0 31.3 30.6 28.1 26.3 27.5 20.0 NA  NA  NA
#> 3  ALP  4919 10364    bota 1 fal 54.5 27.9 31.8 26.0 22.8 25.3 19.5 NA  NA  NA
#> 4  ALP  2211  7105    bota 1 fal 47.6 25.2 25.2 20.6 23.7 17.8   NA NA  NA  NA
#> 5  ALP  1257  7093    bota 1 fal 51.1 24.1  2.7 21.5 17.0   NA   NA NA  NA  NA
#> 6  ALP  4693 10258      oc 1 fal 31.0 10.7 12.4   NA  7.4  8.9  8.0 NA  NA  NA
#>   Dl HmandM3 logBd logBFd logBG logBp logBPC logBT logDd logDl logDp logDPA
#> 1 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 2 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 3 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 4 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 5 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 6 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#>   logGB logGL logGLc logGLP logH logL logLA logLAR logLG logSD logSH logSLC
#> 1    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 2    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 3    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 4    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 5    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 6    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#>   Length Width
#> 1     NA    NA
#> 2     NA    NA
#> 3     NA    NA
#> 4     NA    NA
#> 5     NA    NA
#> 6     NA    NA

## Extract only width with "average" method:
dataExampleWithSummary2 <- CondenseLogs(dataExampleWithLogs,
                               grouping = list(Width = c("Bd", "BT", "Bp", "SD")),
                               method = "average")
head(dataExampleWithSummary2)[, -c(6:20,32:63)]

#>   Site N.inv    UE Especie    Os   GL   Bp   Dp   SD   DD   Bd   Dd BT GLc BFd
#> 1  ALP  3629  8427    avif 1 fal   NA   NA   NA   NA   NA   NA   NA NA  NA  NA
#> 2  ALP  4918 10364    bota 1 fal 54.0 31.3 30.6 28.1 26.3 27.5 20.0 NA  NA  NA
#> 3  ALP  4919 10364    bota 1 fal 54.5 27.9 31.8 26.0 22.8 25.3 19.5 NA  NA  NA
#> 4  ALP  2211  7105    bota 1 fal 47.6 25.2 25.2 20.6 23.7 17.8   NA NA  NA  NA
#> 5  ALP  1257  7093    bota 1 fal 51.1 24.1  2.7 21.5 17.0   NA   NA NA  NA  NA
#> 6  ALP  4693 10258      oc 1 fal 31.0 10.7 12.4   NA  7.4  8.9  8.0 NA  NA  NA
#>   Dl HmandM3 logBd logBFd logBG logBp logBPC logBT logDd logDl logDp logDPA
#> 1 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 2 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 3 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 4 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 5 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 6 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#>   logGB logGL logGLc logGLP logH logL logLA logLAR logLG logSD logSH logSLC
#> 1    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 2    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 3    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 4    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 5    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 6    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#>   Width
#> 1    NA
#> 2    NA
#> 3    NA
#> 4    NA
#> 5    NA
#> 6    NA