Log Ratios of Measurements

Function to compute the log ratios of the measurements relative to standard reference values. By default a reference is provided with the package.

LogRatios(
  data,
  ref = reference$Combi,
  identifiers = c("TAX", "EL"),
  refMeasuresName = "Measure",
  refValuesName = "Standard",
  thesaurusSet = zoologThesaurus,
  joinCategories = NULL,
  mergedMeasures = NULL
)

Arguments

data
A dataframe with the input measurements.

ref
A dataframe including the measurement values used as references. The default ref = reference$Combi provided as package zoolog data.

identifiers
A vector of column names in ref identifying a type of bone. By default identifiers = c("TAX", "EL").

refMeasuresName
The column name in ref identifying the type of bone measurement.

refValuesName
The column name in ref giving the measurement value.

thesaurusSet
A thesaurus allowing datasets with different nomenclatures to be merged. By default thesaurusSet = zoologThesaurus.

joinCategories
A list of named character vectors. Each vector is named by a category in the reference and includes a set of categories in the data for which to compute the log ratios with respect to that reference. When NULL (default) no grouping is considered.

mergedMeasures
A list of character vectors or a single character vector. Each vector identifies a set of measures that the data presents merged in the same column, named as any of them. This practice only makes sense if only one of the measures can appear in each bone element.

Value

A dataframe including the input dataframe and additional columns, one for each extracted log ratio for each relevant measurement in the reference. The name of the added columns are constructed by prefixing each measurement by the internal variable logPrefix.

Details

The identifiers are expected to determine corresponding columns in both data and reference. Each value in these columns identifies the type of bone. By default this is determined by a taxon and a bone element. For any case in the data, the log ratios are computed with respect to the reference values in the same bone type. If the reference does not include that bone type, the corresponding log ratios are set to NA.

For some applications it can be interesting to group some set of bone types into the same reference category to compute the log ratios. The parameter joinCategories allows this grouping. joinCategories must be a list of named vectors, each including the set of categories in the data which should be mapped to the reference category given by its name.

This can be applied to group different species into a single reference species. For instance sheep, capra, and doubtful cases between both (sheep/capra), can be grouped and assigned to the same reference for sheep, by setting joinCategories = list(sheep = c("sheep", "capra", "oc")). Similarly, it can be applied to group different bone elements into a single reference.

Note that the joinCategories option does not remove the distinction between the different bone types in the data, just indicates that for any of them the log ratios must be computed from the same reference.

There are some measures that are restricted to a subset of bones. For instance, GLl is only relevant for the astragalus, while GL is not applicable to it. Thus, there cannot be any ambiguity between both measures since they can be identified by the bone element. This justifies that some users have simplified datasets where a single column records indistinctly GL or GLl. The optional parameter mergedMeasures facilitates the processing of this type of simplified datasets. For the alluded example, mergedMeasures = list(c("GL", "GLl")) automatically selects, for each bone element, the corresponding measure present in the reference.

Observe that if mergedMeasures is set to non mutually exclusive measures, the behaviour is unpredictable.

Examples

## Read an example dataset:
dataFile <- system.file("extdata", "dataValenzuelaLamas2008.csv.gz",
                        package="zoolog")
dataExample <- utils::read.csv2(dataFile,
                                quote = "\"", na = "", header = TRUE,
                                fileEncoding = "UTF-8")
## We can observe the first lines (excluding some columns for visibility):
head(dataExample)[, -c(6:20,32:63)]

#>   Site N.inv    UE Especie    Os   GL   Bp   Dp   SD   DD   Bd   Dd BT GLc BFd
#> 1  ALP  3629  8427    avif 1 fal   NA   NA   NA   NA   NA   NA   NA NA  NA  NA
#> 2  ALP  4918 10364    bota 1 fal 54.0 31.3 30.6 28.1 26.3 27.5 20.0 NA  NA  NA
#> 3  ALP  4919 10364    bota 1 fal 54.5 27.9 31.8 26.0 22.8 25.3 19.5 NA  NA  NA
#> 4  ALP  2211  7105    bota 1 fal 47.6 25.2 25.2 20.6 23.7 17.8   NA NA  NA  NA
#> 5  ALP  1257  7093    bota 1 fal 51.1 24.1  2.7 21.5 17.0   NA   NA NA  NA  NA
#> 6  ALP  4693 10258      oc 1 fal 31.0 10.7 12.4   NA  7.4  8.9  8.0 NA  NA  NA
#>   Dl HmandM3
#> 1 NA      NA
#> 2 NA      NA
#> 3 NA      NA
#> 4 NA      NA
#> 5 NA      NA
#> 6 NA      NA

## We keep now only the first 1000 cases to make the example run sufficiently
## fast. Avoid this step if you want to process the full example dataset.
dataExample <- dataExample[1:1000, ]

## Compute the log-ratios with respect to the default reference in the
## package zoolog:
dataExampleWithLogs <- LogRatios(dataExample)
## The output data frame include new columns with the log-ratios of the
## present measurements, in both data and reference, with a "log" prefix:
head(dataExampleWithLogs)[, -c(6:20,32:63)]

#>   Site N.inv    UE Especie    Os   GL   Bp   Dp   SD   DD   Bd   Dd BT GLc BFd
#> 1  ALP  3629  8427    avif 1 fal   NA   NA   NA   NA   NA   NA   NA NA  NA  NA
#> 2  ALP  4918 10364    bota 1 fal 54.0 31.3 30.6 28.1 26.3 27.5 20.0 NA  NA  NA
#> 3  ALP  4919 10364    bota 1 fal 54.5 27.9 31.8 26.0 22.8 25.3 19.5 NA  NA  NA
#> 4  ALP  2211  7105    bota 1 fal 47.6 25.2 25.2 20.6 23.7 17.8   NA NA  NA  NA
#> 5  ALP  1257  7093    bota 1 fal 51.1 24.1  2.7 21.5 17.0   NA   NA NA  NA  NA
#> 6  ALP  4693 10258      oc 1 fal 31.0 10.7 12.4   NA  7.4  8.9  8.0 NA  NA  NA
#>   Dl HmandM3 logBd logBFd logBG logBp logBPC logBT logDd logDl logDp logDPA
#> 1 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 2 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 3 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 4 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 5 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#> 6 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA    NA     NA
#>   logGB logGL logGLc logGLP logH logL logLA logLAR logLG logSD logSH logSLC
#> 1    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 2    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 3    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 4    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 5    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA
#> 6    NA    NA     NA     NA   NA   NA    NA     NA    NA    NA    NA     NA

## Compute the log-ratios with respect to a different reference:
dataExampleWithLogs2 <- LogRatios(dataExample, ref = reference$Basel)
head(dataExampleWithLogs2)[, -c(6:20,32:63)]

#>   Site N.inv    UE Especie    Os   GL   Bp   Dp   SD   DD   Bd   Dd BT GLc BFd
#> 1  ALP  3629  8427    avif 1 fal   NA   NA   NA   NA   NA   NA   NA NA  NA  NA
#> 2  ALP  4918 10364    bota 1 fal 54.0 31.3 30.6 28.1 26.3 27.5 20.0 NA  NA  NA
#> 3  ALP  4919 10364    bota 1 fal 54.5 27.9 31.8 26.0 22.8 25.3 19.5 NA  NA  NA
#> 4  ALP  2211  7105    bota 1 fal 47.6 25.2 25.2 20.6 23.7 17.8   NA NA  NA  NA
#> 5  ALP  1257  7093    bota 1 fal 51.1 24.1  2.7 21.5 17.0   NA   NA NA  NA  NA
#> 6  ALP  4693 10258      oc 1 fal 31.0 10.7 12.4   NA  7.4  8.9  8.0 NA  NA  NA
#>   Dl HmandM3 logBd logBFd logBG logBp logBPC logBT logDD logDl logDPA logGB
#> 1 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA     NA    NA
#> 2 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA     NA    NA
#> 3 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA     NA    NA
#> 4 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA     NA    NA
#> 5 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA     NA    NA
#> 6 NA      NA    NA     NA    NA    NA     NA    NA    NA    NA     NA    NA
#>   logGL logGLP logLA logLAR logLG logSD logSH logSLC
#> 1    NA     NA    NA     NA    NA    NA    NA     NA
#> 2    NA     NA    NA     NA    NA    NA    NA     NA
#> 3    NA     NA    NA     NA    NA    NA    NA     NA
#> 4    NA     NA    NA     NA    NA    NA    NA     NA
#> 5    NA     NA    NA     NA    NA    NA    NA     NA
#> 6    NA     NA    NA     NA    NA    NA    NA     NA

## Define an altenative reference combining differently the references'
## database:
refComb <- list(cattle = "Nieto", sheep = "Davis", Goat = "Clutton",
                pig = "Albarella", redDeer = "Basel")
userReference <- AssembleReference(refComb)
## Compute the log-ratios with respect to this alternative reference:
dataExampleWithLogs3 <- LogRatios(dataExample, ref = userReference)

## We can be interested in including the first phalanges without
## identification as posterior or anterior, to compute the log ratios with
## respect to the anterior first phalanges.
categoriesP1 <- list('P1 ant' = c("P1 ant", "P1"))
dataExampleWithLogs4 <- LogRatios(dataExample,
                                  joinCategories = categoriesP1)
head(dataExampleWithLogs4)[, -c(6:20,32:63)]

#>   Site N.inv    UE Especie    Os   GL   Bp   Dp   SD   DD   Bd   Dd BT GLc BFd
#> 1  ALP  3629  8427    avif 1 fal   NA   NA   NA   NA   NA   NA   NA NA  NA  NA
#> 2  ALP  4918 10364    bota 1 fal 54.0 31.3 30.6 28.1 26.3 27.5 20.0 NA  NA  NA
#> 3  ALP  4919 10364    bota 1 fal 54.5 27.9 31.8 26.0 22.8 25.3 19.5 NA  NA  NA
#> 4  ALP  2211  7105    bota 1 fal 47.6 25.2 25.2 20.6 23.7 17.8   NA NA  NA  NA
#> 5  ALP  1257  7093    bota 1 fal 51.1 24.1  2.7 21.5 17.0   NA   NA NA  NA  NA
#> 6  ALP  4693 10258      oc 1 fal 31.0 10.7 12.4   NA  7.4  8.9  8.0 NA  NA  NA
#>   Dl HmandM3       logBd logBFd logBG        logBp logBPC logBT        logDd
#> 1 NA      NA          NA     NA    NA           NA     NA    NA           NA
#> 2 NA      NA  0.02435935     NA    NA  0.048386306     NA    NA -0.006466042
#> 3 NA      NA -0.01185283     NA    NA -0.001553828     NA    NA -0.017461427
#> 4 NA      NA -0.16455335     NA    NA -0.045757491     NA    NA           NA
#> 5 NA      NA          NA     NA    NA -0.065140989     NA    NA           NA
#> 6 NA      NA          NA     NA    NA           NA     NA    NA           NA
#>   logDl        logDp logDPA logGB       logGL logGLc logGLP logH logL logLA
#> 1    NA           NA     NA    NA          NA     NA     NA   NA   NA    NA
#> 2    NA  0.008600172     NA    NA -0.01579427     NA     NA   NA   NA    NA
#> 3    NA  0.025305865     NA    NA -0.01179152     NA     NA   NA   NA    NA
#> 4    NA -0.075720714     NA    NA -0.07058107     NA     NA   NA   NA    NA
#> 5    NA -1.045757491     NA    NA -0.03976713     NA     NA   NA   NA    NA
#> 6    NA           NA     NA    NA          NA     NA     NA   NA   NA    NA
#>   logLAR logLG       logSD logSH logSLC
#> 1     NA    NA          NA    NA     NA
#> 2     NA    NA  0.08697848    NA     NA
#> 3     NA    NA  0.05324551    NA     NA
#> 4     NA    NA -0.04786062    NA     NA
#> 5     NA    NA -0.02928938    NA     NA
#> 6     NA    NA          NA    NA     NA