References

Several osteometrical references are provided in zoolog to enable researchers to use the one of their choice. The user can also use their own osteometrical reference if preferred.

reference

referenceSets

referencesDatabase

Format

Each reference is a data.frame including 4 columns:

TAX

The taxon to which each reference bone belongs.

EL

The skeletal element.

Measure

The type of measurement taken on the bone.

Standard

The value of the measurement taken on the bone. All the measurements are expressed in millimetres.

Data Source

Currently, the references include reference values for the main domesticates and their agriotypes (Bos, Ovis, Capra, Sus), and other less frequent species, such as red deer and donkey, drawn from the following publications and resources:

Cattle - Bos

Nieto

Bos taurus. Female cow dated to the Early Bronze Age (Minferri, Catalonia), in Nieto-Espinet (2018) .

Basel

Bos taurus. Inv.nr. 2426 (Hinterwälder; female; 17 years old; live weight: 340 kg; withers height: 113 cm), from Stopp and Deschler-Erb (2018) .

Johnstone

Bos taurus. Standard values from means of cattle measures from Period II (Late Iron Age to Romano-British transition) of Elms Farm, Heybridge (Johnstone and Albarella 2002) .

Degerbøl

Bos primigenius. Female aurochs from Degerbøl and Fredskild (1970) .

Sheep - Ovis

Davis

Ovis aries. Mean values of measurements from a group of adult female Shetland sheep skeletons from a single flock (Davis 1996) .

Clutton

Ovis aries. Mean measurements from a group of male Soay sheep of known age (Clutton-Brock et al. 1990) .

Basel

Ovis musimon. Inv.nr. 2266 (male; adult), from Stopp and Deschler-Erb (2018) .

Uerpmann

Ovis orientalis. Field Museum of Chicago catalogue number: FMC 57951 (female; western Iran) from Uerpmann and Uerpmann (1994) .

Goat - Capra

Basel

Capra hircus. Inv.nr. 1597 (male; adult), from Stopp and Deschler-Erb (2018) .

Clutton

Capra hircus. Mean measurements from a group of goats of unknown age and sex (Clutton-Brock et al. 1990) .

Uerpmann

Capra aegagrus. Measurements based on female and male Capra aegagrus, Natural History Museum in London number: BMNH 651 M and L2 (Taurus Mountains in southern Turkey) from Uerpmann and Uerpmann (1994) .

Pig - Sus

Albarella

Sus domesticus. Mean measurements from a group of Late Neolithic pigs from Durrington Walls, England (Albarella and Payne 2005) .

Basel

Sus scrofa. Inv.nr. 1446 (male; 2-3 years old; life weight: 120 kg) from Stopp and Deschler-Erb (2018) .

Hongo

Sus scrofa. Averaged left and right measurements of a female wild board from near Elaziğ, Turkey. Museum of Comparative Zoology, Harvard University, specimen #51621 (Hongo and Meadow 2000) .

Payne

Sus scrofa. Measurements based on a sample of modern wild boar, Sus scrofa libycus, (male and female; Kizilcahamam, Turkey) from Payne and Bull (1988) , Appendix 2.

Red deer - Cervus

Basel

Cervus elaphus. Inv.nr. 2271 (male; adult) from Stopp and Deschler-Erb (2018) .

Fallow deer - Dama

Haifa

Dama mesopotamica. Adult female modern specimen from Israel (id #1047), curated in Archaeozoology Laboratory at the University of Haifa (Harding and Marom 2021) .

Gazelle - Gazella

Haifa

Gazella gazella. Adult female modern specimen from Israel (id #1037), curated in Archaeozoology Laboratory at the University of Haifa (Harding and Marom 2021) .

Equid - Equus

Haifa

Equus asinus. Adult male modern specimen from Israel (id #1076), curated in Archaeozoology Laboratory at the University of Haifa (Harding and Marom 2021) .

Johnstone

Equus caballus. 3 years old Icelandic mare (all bones fused, female) died in 1961, (Johnstone 2004) . Skeleton held at the Zoologische Staatssammlung Munich in Germany. Specimen ID 1961/29.

European rabbit - Oryctolagus

Nottingham

Oryctolagus cuniculus. Adult male European rabbit from Audley End, Essex, UK, curated in the reference collection at University of Nottingham Arch department (ID RS139) (Ameen 2021) .

The zoolog variable referencesDatabase collects all these references. It is structured as a named list of named lists, following the hierarchy described above:

str(referencesDatabase, max.level = 2)

## List of 14
##  $ Bos taurus           :List of 3
##   ..$ Nieto    :'data.frame':  68 obs. of  4 variables:
##   ..$ Basel    :'data.frame':  50 obs. of  4 variables:
##   ..$ Johnstone:'data.frame':  24 obs. of  4 variables:
##  $ Bos primigenius      :List of 1
##   ..$ Degerbol:'data.frame':  50 obs. of  4 variables:
##  $ Ovis aries           :List of 2
##   ..$ Davis  :'data.frame':  23 obs. of  4 variables:
##   ..$ Clutton:'data.frame':  71 obs. of  4 variables:
##  $ Ovis orientalis      :List of 2
##   ..$ Basel   :'data.frame':  36 obs. of  4 variables:
##   ..$ Uerpmann:'data.frame':  50 obs. of  4 variables:
##  $ Capra hircus         :List of 2
##   ..$ Basel  :'data.frame':  35 obs. of  4 variables:
##   ..$ Clutton:'data.frame':  60 obs. of  4 variables:
##  $ Capra aegagrus       :List of 1
##   ..$ Uerpmann:'data.frame':  50 obs. of  4 variables:
##  $ Sus domesticus       :List of 1
##   ..$ Albarella:'data.frame':  42 obs. of  4 variables:
##  $ Sus scrofa           :List of 3
##   ..$ Basel:'data.frame':  41 obs. of  4 variables:
##   ..$ Hongo:'data.frame':  96 obs. of  4 variables:
##   ..$ Payne:'data.frame':  33 obs. of  4 variables:
##  $ Cervus elaphus       :List of 1
##   ..$ Basel:'data.frame':  14 obs. of  4 variables:
##  $ Dama mesopotamica    :List of 1
##   ..$ Haifa:'data.frame':  60 obs. of  4 variables:
##  $ Gazella gazella      :List of 1
##   ..$ Haifa:'data.frame':  63 obs. of  4 variables:
##  $ Equus asinus         :List of 1
##   ..$ Haifa:'data.frame':  48 obs. of  4 variables:
##  $ Equus caballus       :List of 1
##   ..$ Johnstone:'data.frame':  75 obs. of  4 variables:
##  $ Oryctolagus cuniculus:List of 1
##   ..$ Nottingham:'data.frame':  58 obs. of  4 variables:

Reference Sets

The references' database is organized per taxon. However, in general the zooarchaeological data to be analysed includes several taxa. Thus, the reference dataframe should include one reference standard for each relevant taxon. The zoolog variable referenceSets defines four possible references:

referenceSets

Bos taurus	Bos primigenius	Ovis aries	Ovis orientalis	Capra hircus	Capra aegagrus	Sus domesticus	Sus scrofa	Cervus elaphus	Dama mesopotamica	Gazella gazella	Equus asinus	Equus caballus	Oryctolagus cuniculus

NietoDavisAlbarella	Nieto		Davis				Albarella

Basel	Basel			Basel	Basel			Basel	Basel

Combi	Nieto		Clutton		Clutton			Basel	Basel	Haifa	Haifa	Haifa	Johnstone	Nottingham

Groningen		Degerbol		Uerpmann		Uerpmann		Hongo

Each row defines a reference set consisting of a reference source for each taxon (column). The function AssembleReference allows us to build the reference set taking the selected taxon-specific references from the referencesDatabase.

The zoolog variable reference is a named list including the references defined by referenceSets:

str(reference)

## List of 4
##  $ NietoDavisAlbarella:'data.frame':  133 obs. of  4 variables:
##   ..$ TAX     : Factor w/ 3 levels "bota","ovar",..: 1 1 1 1 1 1 1 1 1 1 ...
##   ..$ EL      : Factor w/ 27 levels "AS","CAL","FE",..: 4 4 4 4 4 4 4 4 4 11 ...
##   ..$ Measure : Factor w/ 24 levels "BFd","BFp","BT",..: 8 9 5 7 12 4 3 11 6 8 ...
##   ..$ Standard: num [1:133] 259 234 78.3 90.2 29 ...
##  $ Basel              :'data.frame':  176 obs. of  4 variables:
##   ..$ TAX     : Factor w/ 5 levels "BOTA","Ovis orientalis",..: 1 1 1 1 1 1 1 1 1 1 ...
##   ..$ EL      : Factor w/ 28 levels "Astragalus","Calcaneus",..: 14 14 14 14 5 5 5 13 13 13 ...
##   ..$ Measure : Factor w/ 26 levels "BFd","BFp","BG",..: 21 13 18 3 5 4 19 6 19 5 ...
##   ..$ Standard: num [1:176] 65.9 83 66.9 58.1 95.3 ...
##  $ Combi              :'data.frame':  558 obs. of  4 variables:
##   ..$ TAX     : Factor w/ 10 levels "bota","OVAR",..: 1 1 1 1 1 1 1 1 1 1 ...
##   ..$ EL      : Factor w/ 56 levels "AS","CAL","FE",..: 4 4 4 4 4 4 4 4 4 11 ...
##   ..$ Measure : Factor w/ 65 levels "BFd","BFp","BT",..: 8 9 5 7 12 4 3 11 6 8 ...
##   ..$ Standard: num [1:558] 259 234 78.3 90.2 29 ...
##  $ Groningen          :'data.frame':  246 obs. of  4 variables:
##   ..$ TAX     : Factor w/ 4 levels "Bos primigenius",..: 1 1 1 1 1 1 1 1 1 1 ...
##   ..$ EL      : Factor w/ 23 levels "Astragalus","Calcaneus",..: 13 13 13 5 5 5 5 5 12 12 ...
##   ..$ Measure : Factor w/ 45 levels "BFp","BG","BT",..: 14 12 2 8 9 4 3 13 8 5 ...
##   ..$ Standard: num [1:246] 69 70 60 359 309 97 89 46 320 100 ...

reference$Combi includes the most comprehensive reference for each species so that more measurements can be considered. It is the default reference for computing the log ratios.

If desired, the user can define their own combinations or can also use their own references, which must be a dataframe with the format described above.

File Structure

referencesDatabase, refereceSets, and reference are exported variables automatically loaded in memory. In addition, zoolog provides in the extdata folder a set of semicolon separated files (csv), generating them:

referenceSets.csv

Defines referenceSets.

referencesDatabase.csv

Defines the structure of referencesDatabase.

...

A csv file for each taxon-specific reference, as named in referencesDatabase.csv.

utils::read.csv2(system.file("extdata", "referencesDatabase.csv",
                             package = "zoolog"))

##                              Genus                 Taxon     Source
## 1                   Cattle - *Bos*            Bos taurus      Nieto
## 2                   Cattle - *Bos*            Bos taurus      Basel
## 3                   Cattle - *Bos*            Bos taurus  Johnstone
## 4                   Cattle - *Bos*       Bos primigenius   Degerbol
## 5                   Sheep - *Ovis*            Ovis aries      Davis
## 6                   Sheep - *Ovis*            Ovis aries    Clutton
## 7                   Sheep - *Ovis*       Ovis orientalis      Basel
## 8                   Sheep - *Ovis*       Ovis orientalis   Uerpmann
## 9                   Goat - *Capra*          Capra hircus      Basel
## 10                  Goat - *Capra*          Capra hircus    Clutton
## 11                  Goat - *Capra*        Capra aegagrus   Uerpmann
## 12                     Pig - *Sus*        Sus domesticus  Albarella
## 13                     Pig - *Sus*            Sus scrofa      Basel
## 14                     Pig - *Sus*            Sus scrofa      Hongo
## 15                     Pig - *Sus*            Sus scrofa      Payne
## 16             Red deer - *Cervus*        Cervus elaphus      Basel
## 17            Fallow deer - *Dama*     Dama mesopotamica      Haifa
## 18             Gazelle - *Gazella*       Gazella gazella      Haifa
## 19                 Equid - *Equus*          Equus asinus      Haifa
## 20                 Equid - *Equus*        Equus caballus  Johnstone
## 21 European rabbit - *Oryctolagus* Oryctolagus cuniculus Nottingham
##                          Filename
## 1       referenceCattle_Nieto.csv
## 2       referenceCattle_Basel.csv
## 3   referenceCattle_Johnstone.csv
## 4    referenceCattle_Degerbol.csv
## 5        referenceSheep_Davis.csv
## 6      referenceSheep_Clutton.csv
## 7        referenceSheep_Basel.csv
## 8     referenceSheep_Uerpmann.csv
## 9         referenceGoat_Basel.csv
## 10      referenceGoat_Clutton.csv
## 11     referenceGoat_Uerpmann.csv
## 12     referencePig_Albarella.csv
## 13         referencePig_Basel.csv
## 14         referencePig_Hongo.csv
## 15         referencePig_Payne.csv
## 16     referenceRedDeer_Basel.csv
## 17        referenceDama_Haifa.csv
## 18     referenceGazelle_Haifa.csv
## 19       referenceEquid_Haifa.csv
## 20   referenceEquid_Johnstone.csv
## 21 referenceRabbit_Nottingham.csv

Acknowledgement

We are grateful to Barbara Stopp and Sabine Deschler-Erb for providing the Basel references for cattle, sheep, goat, wild boar, and red deer (Stopp and Deschler-Erb 2018) , together with the permission to publish them as part of zoolog.

We thank also Francesca Slim and Dimitris Filioglou for providing the references for aurochs, mouflon, wild goat, and wild boar (Degerbøl and Fredskild 1970; Uerpmann and Uerpmann 1994; Hongo and Meadow 2000) in the Groningen set.

We thank Claudia Minniti for providing Johnstone's reference for cattle. (Johnstone and Albarella 2002) .

We are also grateful to Sierra Harding and Nimrod Marom for providing the Haifa standard measurements for donkey, gazelle, and fallow deer (Harding and Marom 2021) .

We thank Carly Ameen and Helene Benkert for providing references for horse (Johnstone 2004) and rabbit (Ameen 2021) .

References

Albarella U, Payne S (2005). “Neolithic pigs from Durrington Walls, Wiltshire, England: a biometrical database.” Journal of Archaeological Science, 32(4), 589--599.

Ameen C (2021). “Measurements from an adult male specimen from Audley End, Essex, UK. in the reference collection at the University of Nottingham Archaeology Department under ID RS139.” Personal communication, included permission to publish them as part of the package zoolog.

Clutton-Brock J, Dennis-Bryan K, Armitage PL, Jewell PA (1990). “Osteology of the Soay sheep.” Bulletin of the British Museum, Natural History. Zoology, 56(1), 1--56.

Davis SJ (1996). “Measurements of a group of adult female Shetland sheep skeletons from a single flock: a baseline for zooarchaeologists.” Journal of archaeological science, 23(4), 593--612.

Degerbøl M, Fredskild B (1970). The Urus (Bos Primigenius Bojanus) and Neolithic Domesticated Cattle (Bos Taurus Domesticus Linné) in Denmark: Zoological and Palynological Investigations, Biologiske skrifter, 17:1. København, (Munksgaard).

Harding S, Marom N (2021). “Measurements compiled for the Zooarchaeology of Southern Phoenicia (ZSP) Project, from the reference collection in the Leon Recanati Institute for Maritime Studies (RIMS, Department of Maritime Civilizations, University of Haifa, Israel).” Personal communication, included permission to publish them as part of the package zoolog.

Hongo H, Meadow RH (2000). “Faunal remains from Prepottery Neolithic levels at Çayönü, southeastern Turkey: a preliminary report focusing on pigs (Sus sp.).” In Archaeozoology of the Near East IVA Proceedings of the fourth international symposium on the archaeozoology of southwestern Asia and adjacent areas. Groningen: ARC Publications, 121--139.

Johnstone C, Albarella U (2002). “The Late Iron Age and Romano-British Mammal and Bird Bone Assemblage from Elms Farm, Heybridge, Essex (Site Code: Hyef93-95).” Technical Report Report 45/2002, tab.16, p. 70, Centre for Archaeology.

Johnstone CJ (2004). A biometric study of equids in the Roman world. Ph.D. thesis, University of York.

Nieto-Espinet A (2018). “Element measure standard biometrical data from a cow dated to the Early Bronze Age (Minferri, Catalonia).” doi: 10.13140/RG.2.2.13512.78081 .

Payne S, Bull G (1988). “Components of variation in measurements of pig bones and teeth, and the use of measurements to distinguish wild from domestic pig remains.” Archaeozoologia, 2(1), 27--66.

Stopp B, Deschler-Erb S (2018). “Measurements compiled from the reference collection in the Integrative Prähistorische und Naturwissenschaftliche Archäologie (IPNA, University of Basel, Switzerland).” Personal communication, included permission to publish them as part of the package zoolog.

Uerpmann M, Uerpmann H (1994). “Animal bone finds from excavation 520 at Qala’at al-Bahrain.” In Hojlund F, Andersen HH (eds.), Gala'at Al-Bahrain. 1 The Northern City Wall And The Islamic Fortress, 417--444. Jutland Archaeological Society.