References

Several osteometrical references are provided in zoolog to enable researchers to use the one of their choice. The user can also use their own osteometrical reference if preferred.

reference

referenceSets

referencesDatabase

Format

Each reference is a data.frame including 4 columns:

TAX

The taxon to which each reference bone belongs.

EL

The skeletal element.

Measure

The type of measurement taken on the bone.

Standard

The value of the measurement taken on the bone. All the measurements are expressed in millimetres.

Data Source

Currently, the references include reference values for the main domesticates and their agriotypes (Bos, Ovis, Capra, Sus), and reed deer (Cervus elaphus) originated from the following publications and resources:

Cattle - Bos

Nieto

Bos taurus. Female cow dated to the Early Bronze Age (Minferri, Catalonia), in Nieto-Espinet (2018) .

Basel

Bos taurus Inv.nr. 2426 (Hinterwälder; female; 17 years old; live weight: 340 kg; withers height: 113 cm), from Stopp and Deschler-Erb (2018) .

Degerbøl

Bos primigenius. Female aurochs from Degerbøl and Fredskild (1970) .

Sheep - Ovis

Davis

Ovis aries. Mean values of measurements from a group of adult female Shetland sheep skeletons from a single flock (Davis 1996) .

Basel

Ovis musimon Inv.nr. 2266 (male; adult), from Stopp and Deschler-Erb (2018) .

Clutton

Ovis aries. Mean measurements from a group of Soay known-age sheep males (Clutton-Brock et al. 1990) .

Uerpmann

Ovis orientalis Field Museum of Chicago catalogue number: FMC 57951 (female; western Iran) from Uerpmann and Uerpmann (1994) .

Goat - Capra

Basel

Capra hircus Inv.nr. 1597 (male; adult), from Stopp and Deschler-Erb (2018) .

Clutton

Capra hircus. Mean measurements from a group of goats of unknown age and sex (Clutton-Brock et al. 1990) .

Uerpmann

Measurements based on female and male Capra aegagrus, Natural History Museum in London number: BMNH 651 M and L2 (Taurus Mountains in southern Turkey) from Uerpmann and Uerpmann (1994) .

Pig - Sus

Albarella

Sus domesticus Mean measurements from a group of Late Neolithic pigs from Durrington Walls, England (Albarella and Payne 2005) .

Basel

Sus scrofa Inv.nr. 1446 (male; 2-3 years old; life weight: 120 kg) from Stopp and Deschler-Erb (2018) .

Hongo

Sus scrofa. Averaged left and right measurements of a female wild board from near Elaziğ, Turkey. Museum of Comparative Zoology, Harvard University, specimen #51621 (Hongo and Meadow 2000) .

Payne

Measurements based on a sample of modern wild boar, Sus scrofa libycus, (male and female; Kizilcahamam, Turkey) from Payne and Bull (1988) , Appendix 2.

Red deer - Cervus elaphus

Basel

Cervus elaphus Inv.nr. 2271 (male; adult) from Stopp and Deschler-Erb (2018) .

The zoolog variable referencesDatabase collects all these references. It is structured as a named list of named lists, following the hierarchy described above:

str(referencesDatabase, max.level = 2)

## List of 9
##  $ Bos taurus     :List of 2
##   ..$ Nieto:'data.frame':  68 obs. of  4 variables:
##   ..$ Basel:'data.frame':  50 obs. of  4 variables:
##  $ Bos primigenius:List of 1
##   ..$ Degerbol:'data.frame':  50 obs. of  4 variables:
##  $ Ovis aries     :List of 2
##   ..$ Davis  :'data.frame':  23 obs. of  4 variables:
##   ..$ Clutton:'data.frame':  71 obs. of  4 variables:
##  $ Ovis orientalis:List of 2
##   ..$ Basel   :'data.frame':  36 obs. of  4 variables:
##   ..$ Uerpmann:'data.frame':  50 obs. of  4 variables:
##  $ Capra hircus   :List of 2
##   ..$ Basel  :'data.frame':  35 obs. of  4 variables:
##   ..$ Clutton:'data.frame':  60 obs. of  4 variables:
##  $ Capra aegagrus :List of 1
##   ..$ Uerpmann:'data.frame':  50 obs. of  4 variables:
##  $ Sus domesticus :List of 1
##   ..$ Albarella:'data.frame':  11 obs. of  4 variables:
##  $ Sus scrofa     :List of 3
##   ..$ Basel:'data.frame':  41 obs. of  4 variables:
##   ..$ Hongo:'data.frame':  96 obs. of  4 variables:
##   ..$ Payne:'data.frame':  33 obs. of  4 variables:
##  $ Cervus elaphus :List of 1
##   ..$ Basel:'data.frame':  14 obs. of  4 variables:

Reference Sets

The references' database is organized per taxon. However, in general the data to be analysed includes several taxa. Thus, the reference dataframe should include one reference specimen for each relevant taxon. The zoolog variable referenceSets defines four possible references:

referenceSets

Bos taurus	Bos primigenius	Ovis aries	Ovis orientalis	Capra hircus	Capra aegagrus	Sus domesticus	Sus scrofa	Cervus elaphus

NietoDavisAlbarella	Nieto		Davis				Albarella

Basel	Basel			Basel	Basel			Basel	Basel

Combi	Nieto		Clutton		Clutton			Basel	Basel

Groningen		Degerbol		Uerpmann		Uerpmann		Hongo

Each row defines a reference consisting in a reference source for each taxon (column). The function AssembleReference allows us to build the reference set taking the selected taxon-specific references from the referencesDatabase.

The zoolog variable reference is a named list including the references defined by referenceSets:

str(reference)

## List of 4
##  $ NietoDavisAlbarella:'data.frame':  102 obs. of  4 variables:
##   ..$ TAX     : Factor w/ 3 levels "bota","ovar",..: 1 1 1 1 1 1 1 1 1 1 ...
##   ..$ EL      : Factor w/ 12 levels "AS","CAL","FE",..: 4 4 4 4 4 4 4 4 4 11 ...
##   ..$ Measure : Factor w/ 13 levels "BFd","BFp","BT",..: 8 9 5 7 12 4 3 11 6 8 ...
##   ..$ Standard: num [1:102] 259 234 78.3 90.2 29 ...
##  $ Basel              :'data.frame':  176 obs. of  4 variables:
##   ..$ TAX     : Factor w/ 5 levels "BOTA","Ovis orientalis",..: 1 1 1 1 1 1 1 1 1 1 ...
##   ..$ EL      : Factor w/ 28 levels "Astragalus","Calcaneus",..: 14 14 14 14 5 5 5 13 13 13 ...
##   ..$ Measure : Factor w/ 27 levels "BFd","BFp","BG",..: 21 13 18 3 5 4 19 6 19 5 ...
##   ..$ Standard: num [1:176] 65.9 83 66.9 58.1 95.3 ...
##  $ Combi              :'data.frame':  254 obs. of  4 variables:
##   ..$ TAX     : Factor w/ 5 levels "bota","OVAR",..: 1 1 1 1 1 1 1 1 1 1 ...
##   ..$ EL      : Factor w/ 38 levels "AS","CAL","FE",..: 4 4 4 4 4 4 4 4 4 11 ...
##   ..$ Measure : Factor w/ 31 levels "BFd","BFp","BT",..: 8 9 5 7 12 4 3 11 6 8 ...
##   ..$ Standard: num [1:254] 259 234 78.3 90.2 29 ...
##  $ Groningen          :'data.frame':  246 obs. of  4 variables:
##   ..$ TAX     : Factor w/ 4 levels "Bos primigenius",..: 1 1 1 1 1 1 1 1 1 1 ...
##   ..$ EL      : Factor w/ 23 levels "Astragalus","Calcaneus",..: 13 13 13 5 5 5 5 5 12 12 ...
##   ..$ Measure : Factor w/ 45 levels "BFp","BG","BT",..: 14 12 2 8 9 4 3 13 8 5 ...
##   ..$ Standard: num [1:246] 69 70 60 359 309 97 89 46 320 100 ...

reference$Combi includes the most comprehensive reference for each species so that more measurements can be considered. It is the default reference for computing the log ratios.

If desired, the user can define their own combinations or can also use their own references, which must be a dataframe with the format described above.

File Structure

referencesDatabase, refereceSets, and reference are exported variables automatically loaded in memory. In addition, zoolog provides in the extdata folder a set of semicolon separated files (csv), generating them:

referenceSets.csv

Defines referenceSets.

referencesDatabase.csv

Defines the structure of referencesDatabase.

...

A csv file for each taxon-specific reference, as named in referencesDatabase.csv.

utils::read.csv2(system.file("extdata", "referencesDatabase.csv",
                             package = "zoolog"))

##              Taxon    Source                     Filename
## 1       Bos taurus     Nieto    referenceCattle_Nieto.csv
## 2       Bos taurus     Basel    referenceCattle_Basel.csv
## 3  Bos primigenius  Degerbol referenceCattle_Degerbol.csv
## 4       Ovis aries     Davis     referenceSheep_Davis.csv
## 5  Ovis orientalis     Basel     referenceSheep_Basel.csv
## 6       Ovis aries   Clutton   referenceSheep_Clutton.csv
## 7  Ovis orientalis  Uerpmann  referenceSheep_Uerpmann.csv
## 8     Capra hircus     Basel      referenceGoat_Basel.csv
## 9     Capra hircus   Clutton    referenceGoat_Clutton.csv
## 10  Capra aegagrus  Uerpmann   referenceGoat_Uerpmann.csv
## 11  Sus domesticus Albarella   referencePig_Albarella.csv
## 12      Sus scrofa     Basel       referencePig_Basel.csv
## 13      Sus scrofa     Hongo       referencePig_Hongo.csv
## 14      Sus scrofa     Payne       referencePig_Payne.csv
## 15  Cervus elaphus     Basel   referenceRedDeer_Basel.csv

Acknowledgement

We are grateful to Barbara Stopp and Sabine Deschler-Erb for providing the Basel references (Stopp and Deschler-Erb 2018) together with the permission to publish them as part of zoolog.

We thank also to Francesca Slim and Dimitris Filioglou for providing the Groningen references (Degerbøl and Fredskild 1970; Uerpmann and Uerpmann 1994; Hongo and Meadow 2000) .

References

Albarella U, Payne S (2005). “Neolithic pigs from Durrington Walls, Wiltshire, England: a biometrical database.” Journal of Archaeological Science, 32(4), 589--599.

Clutton-Brock J, Dennis-Bryan K, Armitage PL, Jewell PA (1990). “Osteology of the Soay sheep.” Bulletin of the British Museum, Natural History. Zoology, 56(1), 1--56.

Davis SJ (1996). “Measurements of a group of adult female Shetland sheep skeletons from a single flock: a baseline for zooarchaeologists.” Journal of archaeological science, 23(4), 593--612.

Degerbøl M, Fredskild B (1970). The Urus (Bos Primigenius Bojanus) and Neolithic Domesticated Cattle (Bos Taurus Domesticus Linné) in Denmark: Zoological and Palynological Investigations, Biologiske skrifter, 17:1. København, (Munksgaard).

Hongo H, Meadow RH (2000). “Faunal remains from Prepottery Neolithic levels at Çayönü, southeastern Turkey: a preliminary report focusing on pigs (Sus sp.).” In Archaeozoology of the Near East IVA Proceedings of the fourth international symposium on the archaeozoology of southwestern Asia and adjacent areas. Groningen: ARC Publications, 121--139.

Nieto-Espinet A (2018). “Element measure standard biometrical data from a cow dated to the Early Bronze Age (Minferri, Catalonia).” doi: 10.13140/RG.2.2.13512.78081 , https://www.researchgate.net/publication/326010953.

Payne S, Bull G (1988). “Components of variation in measurements of pig bones and teeth, and the use of measurements to distinguish wild from domestic pig remains.” Archaeozoologia, 2(1), 27--66.

Stopp B, Deschler-Erb S (2018). “Measurements compiled from the reference collection in the Integrative Prähistorische und Naturwissenschaftliche Archäologie (IPNA, University of Basel, Switzerland).” Personal communication, included permission to publish them as part of the package zoolog.

Uerpmann M, Uerpmann H (1994). “Animal bone finds from excavation 520 at Qala’at al-Bahrain.” In Hojlund F, Andersen HH (eds.), Gala'at Al-Bahrain. 1 The Northern City Wall And The Islamic Fortress, 417--444. Jutland Archaeological Society.